requirements.txt
Pillow
pytz
inky